Moved styled-components of Match components into Match directories
