+++ src/components/Matches/MatchPresentational/Match.js
@@ -1,10 +1,10 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { CardStyle } from "../../styles/Cards";
-import { UserLike } from "../Helpers";
-import { ImgWithStatusText } from "../Helpers";
-import noImage from "../../resources/img/no-image.png";
+import CardStyle from "./MatchStyle";
+import { UserLike } from "../../Helpers";
+import { ImgWithStatusText } from "../../Helpers";
+import noImage from "../../../resources/img/no-image.png";
 
 function Match(props) {
   const { property, toggleModal, isLiked, toggleLikeHandler } = props;
