--- src/styles/Cards/index.js
@@ -1,5 +1,0 @@
-import CardsContainer from "./CardContainerStyle";
-import Card from "./Card";
-
-export const CardsContainerStyle = CardsContainer;
-export const CardStyle = Card;
